+++ src/views/DocumentSearch.vue
@@ -4,6 +4,7 @@ import type {
   AmountFilterState,
   TimeFilterState,
   StringFilterState,
+  IndexFilterState,
   SizeFilterState,
   FiltersState,
   ClientQuery,
@@ -12,6 +13,7 @@ import type {
   AmountSearchResult,
   TimeSearchResult,
   StringSearchResult,
+  IndexSearchResult,
   SizeSearchResult,
 } from "@/types"
 
@@ -22,7 +24,8 @@ import RelFiltersResult from "@/components/RelFiltersResult.vue"
 import AmountFiltersResult from "@/components/AmountFiltersResult.vue"
 import TimeFiltersResult from "@/components/TimeFiltersResult.vue"
 import StringFiltersResult from "@/components/StringFiltersResult.vue"
-import SizeFilterResult from "@/components/SizeFiltersResult.vue"
+import IndexFiltersResult from "@/components/IndexFiltersResult.vue"
+import SizeFiltersResult from "@/components/SizeFiltersResult.vue"
 import NavBar from "@/components/NavBar.vue"
 import Footer from "@/components/Footer.vue"
 import Button from "@/components/Button.vue"
@@ -137,7 +140,7 @@ onBeforeUnmount(() => {
 
 const updateFiltersProgress = ref(0)
 // A non-read-only version of filters state so that we can modify it as necessary.
-const filtersState = ref<FiltersState>({ rel: {}, amount: {}, time: {}, str: {}, size: null })
+const filtersState = ref<FiltersState>({ rel: {}, amount: {}, time: {}, str: {}, index: [], size: null })
 // We keep it in sync with upstream version.
 watchEffect((onCleanup) => {
   // We copy to make a read-only value mutable.
@@ -186,6 +189,17 @@ async function onStringFiltersStateUpdate(id: string, s: StringFilterState) {
     const updatedState = { ...filtersState.value }
     updatedState.str = { ...updatedState.str }
     updatedState.str[id] = s
+    await postFilters(router, route, updatedState, updateFiltersProgress)
+  } finally {
+    updateFiltersProgress.value -= 1
+  }
+}
+
+async function onIndexFiltersStateUpdate(s: IndexFilterState) {
+  updateFiltersProgress.value += 1
+  try {
+    const updatedState = { ...filtersState.value }
+    updatedState.index = s
     await postFilters(router, route, updatedState, updateFiltersProgress)
   } finally {
     updateFiltersProgress.value -= 1
@@ -296,7 +310,15 @@ const filtersEnabled = ref(false)
             :update-progress="updateFiltersProgress"
             @update:state="onStringFiltersStateUpdate(doc._id, $event)"
           />
-          <SizeFilterResult
+          <IndexFiltersResult
+            v-if="doc._type === 'index'"
+            :search-total="searchTotal"
+            :result="doc as IndexSearchResult"
+            :state="filtersState.index"
+            :update-progress="updateFiltersProgress"
+            @update:state="onIndexFiltersStateUpdate($event)"
+          />
+          <SizeFiltersResult
             v-if="doc._type === 'size'"
             :search-total="searchTotal"
             :result="doc as SizeSearchResult"